Update TaggedUnion: Add Tag
