+++ src/PrimeFuncPack.Core/PrimeFuncPack.Core.TaggedUnion/TaggedUnion.Fold.cs
@@ -1,0 +1,18 @@
+﻿#nullable enable
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace System
+{
+    partial struct TaggedUnion<TTag, TFirst, TSecond>
+    {
+        //public static Optional<TResult> Fold<TResult>(
+        //    in Func<TFirst, TResult> mapFirst,
+        //    in Func<TSecond, TResult> mapSecond)
+        //    =>
+    }
+}
