+++ src/PrimeFuncPack.Core/PrimeFuncPack.Core.TaggedUnion/TaggedUnion.cs
@@ -2,7 +2,9 @@
 
 namespace System
 {
-    public readonly partial struct TaggedUnion<TFirst, TSecond>
+    public readonly partial struct TaggedUnion<TTag, TFirst, TSecond>
+        where TFirst : TTag
+        where TSecond : TTag
     {
         private readonly Box<TFirst>? boxFirst;
 
@@ -13,6 +15,16 @@ namespace System
         public bool IsFirst => boxFirst is object;
 
         public bool IsSecond => boxSecond is object;
+
+        public Optional<TTag> GetTag()
+        {
+            var @this = this;
+
+            return
+                default(Optional<TTag>)
+                .Or(() => @this.TryGetFirst().Map<TTag>(value => value))
+                .Or(() => @this.TryGetSecond().Map<TTag>(value => value));
+        }
 
         public Optional<TFirst> TryGetFirst() => boxFirst switch
         {
